Extracted time_info_data into its own source files
